types: move decision and diagnostic error method related tests to types package
Addresses IDX-50
